+++ main.tf
@@ -13,6 +13,7 @@ provider "aws" {
 }
 
 resource "aws_security_group" "cog_sg" {
+  vpc_id = aws_vpc.cog_vpc.id
   name = "COG SG"
 
   ingress {
@@ -45,8 +46,10 @@ resource "aws_key_pair" "wkl-cognixus-key" {
 resource "aws_instance" "gitea-ec2" {
   ami           = "ami-060e277c0d4cce553" #Ubuntu Server 24.04 LTS
   instance_type = "t2.micro"
+  subnet_id     = aws_subnet.cog_pub_subnet.id
   key_name      = aws_key_pair.wkl-cognixus-key.key_name
-  security_groups = [aws_security_group.cog_sg.name]
+  #security_groups = [aws_security_group.cog_sg.name]
+  vpc_security_group_ids = [aws_security_group.cog_sg.id]
 
   user_data = file("install_docker.sh")