+++ changelog.md
@@ -14,3 +14,4 @@
 - Install Docker and Docker Compose.
 - Install Gitea and Postgres.
 - Expose port 80.
+- Create VPC with single subnet.